sound_classifier_system/requirements.txt
librosa
numpy
sounddevice
joblib
python-dotenv
uvicorn
fastapi
yt_dlp
pytube
scipy
soundata
pydub
pyrubberband
noisereduce
pandas
scikit-learn
soundfile
matplotlib
python-multipart
websockets
psutil
